moved 'foo_state' to separate file
