+++ mission/finite_state_machine/src/sm_classes/__init__.py
@@ -1,1 +1,2 @@
 from gate_search_state import GateSearchState
+from foo_state import FooState
